Add QStyleHints::keyboardAutoRepeatRateF()
The keyboardAutoRepeatRate() function returning an int assumes
the rate can not be lower than one event per second, but this
is not always the case, e.g. on macOS where the slowest setting
for key repeat results in a two second delay between events.

+++ src/widgets/widgets/qabstractspinbox_p.h
@@ -102,7 +102,7 @@ public:
     int spinClickTimerInterval = 100;
     int spinClickThresholdTimerId = -1;
     int spinClickThresholdTimerInterval = -1;
-    int effectiveSpinRepeatRate = 1;
+    qreal effectiveSpinRepeatRate = 1;
     int acceleration = 0;
     int wheelDeltaRemainder = 0;
